+++ Sources/KeyboardKit/Previews/KeyboardPreviews+Themes.swift
@@ -1,0 +1,14 @@
+//
+//  KeyboardPreviews+Themes.swift
+//  KeyboardKit
+//
+//  Created by Daniel Saidi on 2021-03-25.
+//  Copyright © 2021-2024 Daniel Saidi. All rights reserved.
+//
+
+import SwiftUI
+
+public extension KeyboardThemeContext {
+
+    static var preview: KeyboardThemeContext { .init() }
+}